update airbyte-ci bump-version to push more minimal changes (#34339)
I noticed that `airbyte-ci connectors bump-version ` changes more things than necessary. In particular, it reserialize json, which means we were losing comments and reordering the various keys. I moved to doing a search-and-replace. In addition, I also noticed that the changelog changes could be slightly improved: I'm not increasing the space for the version (so we can accomodate 0.21.53, or 0.125.2),  and adding a new line at the end of the file (which was flagged as a difference by `diff`